app/requirements.txt
venv used for this project:

Django	4.2.1
ExifRead	3.0.0
Jinja2	3.1.2
MarkupSafe	2.1.2
Pillow	9.5.0
PyJWT	2.7.0
asgiref	3.6.0
astroid	2.15.5
certifi	2023.5.7
cffi	1.15.1
charset-normalizer	3.1.0
click	8.1.3
colorama	0.4.6
coreapi	2.3.3
coreschema	0.0.4
cryptography	40.0.2
defusedxml	0.7.1
dill	0.3.6
django-templated-mail	1.1.1
djangorestframework	3.14.0
djangorestframework-simplejwt	5.2.2
djoser	2.2.0
drf-yasg	1.21.5
exceptiongroup	1.1.1
filesystemlib	0.1.0
idna	3.4
inflection	0.5.1
iniconfig	2.0.0
isort	5.12.0
itypes	1.2.0
lazy-object-proxy	1.9.0
mccabe	0.7.0
oauthlib	3.2.2
packaging	23.1
pip	22.2.2
platformdirs	3.5.1
pluggy	1.0.0
pycparser	2.21
pylint	2.17.4
pytest	7.3.1
pytest-mock	3.10.0
python3-openid	3.2.0
pytz	2023.3
requests	2.30.0
requests-oauthlib	1.3.1
ruamel.yaml	0.17.26
ruamel.yaml.clib	0.2.7
setuptools	63.2.0
social-auth-app-django	5.2.0
social-auth-core	4.4.2
sqlparse	0.4.4
tomli	2.0.1
tomlkit	0.11.8
typing-extensions	4.5.0
tzdata	2023.3
uritemplate	4.1.1
urllib3	2.0.2
wrapt	1.15.0